backend/requirements.txt
Django
python-decouple
djangorestframework
django-cors-headers
requests